Move quat-euler conversions to math/ and export from the top-level barrel.
